fixing python script

+++ scripts/pig_game.py
@@ -1,8 +1,8 @@
 import subprocess
 
 try:
-    subprocess.run(["kubectl delete -f k8s/pigGame/pig_game_deploy.yaml"])
-    subprocess.run(["kubectl delete -f k8s/pigGame/pig_game_service.yaml"])
+    subprocess.run(["kubectl delete -f k8s/pigGame/pig_game_deploy.yaml"]) || subprocess.run(["kubectl create -f k8s/pigGame/pig_game_deploy.yaml"])
+    
 except:
-     subprocess.run(["kubectl create -f k8s/pigGame/pig_game_deploy.yaml"])
-     subprocess.run(["kubectl create -f k8s/pigGame/pig_game_service.yaml"])
+      subprocess.run(["kubectl create -f k8s/pigGame/pig_game_service.yaml"]) || subprocess.run(["kubectl delete -f k8s/pigGame/pig_game_service.yaml"])
+
